Conflicts:
	libguile/gc_os_dep.c

+++ libguile/inet_aton.c
@@ -103,14 +103,14 @@ inet_aton(const char *cp_arg, struct in_addr *addr)
 				base = 8;
 		}
 		while ((c = *cp) != '\0') {
-			if (isascii(c) && isdigit(c)) {
+                        if (isascii ((int) c) && isdigit ((int) c)) {
 				val = (val * base) + (c - '0');
 				cp++;
 				continue;
 			}
-			if (base == 16 && isascii(c) && isxdigit(c)) {
+                        if (base == 16 && isascii ((int) c) && isxdigit ((int) c)) {
 				val = (val << 4) + 
-					(c + 10 - (islower(c) ? 'a' : 'A'));
+                                  (c + 10 - (islower((int) c) ? 'a' : 'A'));
 				cp++;
 				continue;
 			}
@@ -132,7 +132,7 @@ inet_aton(const char *cp_arg, struct in_addr *addr)
 	/*
 	 * Check for trailing characters.
 	 */
-	if (*cp && (!isascii(*cp) || !isspace(*cp)))
+	if (*cp && (!isascii ((int) (*cp)) || !isspace ((int) (*cp))))
 		return (0);
 	/*
 	 * Concoct the address according to
